+++ source/config.cpp
@@ -176,7 +176,6 @@ void Config::save() {
     j["randomTID"] = this->randomTID;
     j["customTitle"] = this->customTitle;
     j["forceInstall"] = this->forceInstall;
-    j["show3dsRoms"] = this->show3dsRoms;
     j["artNotify"] = this->artNotify;
     j["manageIcons"] = this->manageIcons;
     j["gbaScreen"] = this->gbaScreen;
@@ -197,7 +196,6 @@ void Config::load() {
     if (j.contains("randomTID")) this->randomTID = j["randomTID"];
     if (j.contains("customTitle")) this->customTitle = j["customTitle"];
     if (j.contains("forceInstall")) this->forceInstall = j["forceInstall"];
-    if (j.contains("show3dsRoms")) this->show3dsRoms = j["show3dsRoms"];
     if (j.contains("artNotify")) this->artNotify = j["artNotify"];
     if (j.contains("manageIcons")) this->manageIcons = j["manageIcons"];
     if (j.contains("gbaScreen")) this->gbaScreen = j["gbaScreen"];